% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mysplom2.R
\name{corplot}
\alias{corplot}
\title{corplot}
\usage{
corplot(X, main = "", labels = NULL, file = "")
}
\arguments{
\item{X}{a matrix whose columns refer to parameters and whose rows to samples}

\item{main}{the plot title (if desired)}

\item{labels}{a character vector of length \code{ncol(X)} to label parameters}

\item{file}{a filename to save pdf output plot to (else plotted to default to device)}
}
\value{
None
}
\description{
A hexbinned pairs and density plot to examine correlations.
}
\details{
The diagonal panels are histograms of each parameter representing estimates of marginal density. The lower half shows hexbinned plots of correlations between parameters (estimates of pairwise joint densities) together with a loess smoother to indicate trend.
}
\examples{
corplot(matrix(rnorm(3e4),ncol=3),labels=c('x','y','z'),main='3D isotropic Gaussian')
}
\author{
Pete Dodd
}

