% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AIES.R
\name{getAIESamples}
\alias{getAIESamples}
\title{getAIESamples}
\usage{
getAIESamples(halfN, nosteps, lnprob, x0, a = 2, eps = 0.01, CPUs = 1,
  dynlib = NULL, echo = FALSE, lgf = NULL)
}
\arguments{
\item{halfN}{half the number of walkers in the ensemble}

\item{nosteps}{the number of steps taken (chainlengths)}

\item{lnprob}{the loglikelihood function}

\item{x0}{a vector of initial parameter values}

\item{a}{alg tuning: todo}

\item{eps}{alg tuning: todo}

\item{CPUs}{number of CPUs to use (experimental; not working)}

\item{dynlib}{if a compiled library is need to evaluate \code{lnprob}}

\item{echo}{if \code{TRUE} then more verbose output}

\item{lgf}{if present, the filename for a logfile to dump information into}
}
\value{
\code{list(chains=list(R1=R2,R2=R2),acc=acc,accv=accv,a=a, llike=list(LL1=LL1,LL2=LL2),lnprob=lnprob)} - returned in this form to allow continuation with \code{continueAIESamples}. The chains \code{R1} and \code{R2} are matrices with columns for parameteres and rows for iterations.
}
\description{
Function to obtain affine-invariant ensemble samples
}
\details{
todo: provenance and more on arguments
}
\examples{
#Rosenrbock banana function
rosen <- function(x) -(1-x[1])^2 - 100*(x[2] - x[1]^2)^2

#50 walkers for 200 steps
S <- getAIESamples(50,200,arosen,runif(2))
plotAIEchains(S)
SP <- processAIESamples(S,burnin=50)
corplot(SP)

#continue for another 100 steps
S <- continueAIESamples(Y=S,nosteps=100)
}
\author{
Pete Dodd
}

